tools/obd2_vehicle/requirements.txt
# OBD2 Vehicle Simulator - Dependencies
# Python 3.8+ required

# CAN bus interface
python-can>=4.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: for development
# black>=23.0.0
# mypy>=1.0.0
